Move App inside a folder
